Day 9 Part 2 WIP

+++ AdventOfCodeSolutions/include/day8.h
@@ -203,7 +203,7 @@ void day8()
 		//Find size
 		if (firstLine == false)
 		{
-			width = line.length();
+			width = static_cast<int>(line.length());
 			firstLine = true;
 		}
